Moving fonts definitions to a separate directory
